Migrate SendDailyNotifications Task -> Command
