+++ tests/IntegrationTests/Tasks/SendDailyNotificationsTask/WeeklySummaryGetterIntegrationTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Olz\Tests\IntegrationTests\Tasks\SendDailyNotificationsTask;
+namespace Olz\Tests\IntegrationTests\Command\SendDailyNotificationsCommand;
 
+use Olz\Command\SendDailyNotificationsCommand\WeeklySummaryGetter;
 use Olz\Entity\User;
-use Olz\Tasks\SendDailyNotificationsTask\WeeklySummaryGetter;
 use Olz\Tests\Fake;
 use Olz\Tests\IntegrationTests\Common\IntegrationTestCase;
 use Olz\Utils\DbUtils;
@@ -15,7 +15,7 @@ use Olz\Utils\FixedDateUtils;
 /**
  * @internal
  *
- * @covers \Olz\Tasks\SendDailyNotificationsTask\WeeklySummaryGetter
+ * @covers \Olz\Command\SendDailyNotificationsCommand\WeeklySummaryGetter
  */
 final class WeeklySummaryGetterIntegrationTest extends IntegrationTestCase {
     public function testWeeklySummaryGetter(): void {
